core: Run DKG stuffs only if the node is in DKG set (#302)

+++ core/test/fake-transport.go
@@ -1,6 +1,23 @@
 // Copyright 2018 The dexon-consensus Authors
 // This file is part of the dexon-consensus library.
 //
+// The dexon-consensus library is free software: you can redistribute it
+// and/or modify it under the terms of the GNU Lesser General Public License as
+// published by the Free Software Foundation, either version 3 of the License,
+// or (at your option) any later version.
+//
+// The dexon-consensus library is distributed in the hope that it will be
+// useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser
+// General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the dexon-consensus library. If not, see
+// <http://www.gnu.org/licenses/>.
+
+// Copyright 2018 The dexon-consensus Authors
+// This file is part of the dexon-consensus library.
+//
 // The dexon-consensus library is free software: you can redistribute it and/or
 // modify it under the terms of the GNU Lesser General Public License as
 // published by the Free Software Foundation, either version 3 of the License,
@@ -161,7 +178,7 @@ func (t *FakeTransport) Host() (chan *TransportEnvelope, error) {
 }
 
 // WaitForPeers implements TransportServer.WaitForPeers method.
-func (t *FakeTransport) WaitForPeers(numPeers int) (err error) {
+func (t *FakeTransport) WaitForPeers(numPeers uint32) (err error) {
 	t.peers = make(map[types.NodeID]fakePeerRecord)
 	for {
 		envelope := <-t.recvChannel
@@ -172,7 +189,7 @@ func (t *FakeTransport) WaitForPeers(numPeers int) (err error) {
 			sendChannel: newPeer.recvChannel,
 			pubKey:      newPeer.pubKey,
 		}
-		if len(t.peers) == numPeers {
+		if uint32(len(t.peers)) == numPeers {
 			break
 		}
 	}
